@@ -3,7 +3,7 @@ description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -18,35 +18,33 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.2 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  http: ^0.13.0
+  http: ^1.2.1
   shared_preferences: ^2.0.5
-  rxdart: ^0.26.0
+  rxdart: ^0.27.7
   fluttertoast: ^8.0.3
   sqflite: ^2.0.0+3
   path_provider: ^2.0.1
-  font_awesome_flutter: ^9.0.0
-  carousel_slider: ^2.1.0
-  lottie: ^1.0.1
-  age: ^0.0.1
-  advance_pdf_viewer: ^2.0.0
-  keyboard_actions: ^3.4.0 
+  font_awesome_flutter: ^10.7.0
+  carousel_slider: ^4.2.1
+  lottie: ^3.1.0
+  pdfx: ^2.6.0
+  keyboard_actions: ^4.2.0
   #syncfusion_flutter_charts: ^18.4.44+1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_launcher_icons: ^0.9.0
+  flutter_launcher_icons: ^0.13.1
 
 flutter_icons:
   android: "launcher_icon"
@@ -59,7 +57,6 @@ flutter_icons:
 
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
