@@ -7,15 +7,20 @@ class BeneficiariosBloc {
   final beneficiariosDatabase = BeneficiariosDatabase();
   final beneficiariosApi = BeneficiariosApi();
 
-  final _beneficiariosController = BehaviorSubject<List<BeneficiariosGeneral>>();
-  final _beneficiariosAfiliadosController = BehaviorSubject<List<BeneficiariosGeneral>>();
+  final _beneficiariosController =
+      BehaviorSubject<List<BeneficiariosGeneral>>();
+  final _beneficiariosAfiliadosController =
+      BehaviorSubject<List<BeneficiariosGeneral>>();
 
-  Stream<List<BeneficiariosGeneral>> get beneficiariosControllerStream => _beneficiariosController.stream;
-  Stream<List<BeneficiariosGeneral>> get beneficiariosAfiliadosControllerStream => _beneficiariosAfiliadosController.stream;
+  Stream<List<BeneficiariosGeneral>> get beneficiariosControllerStream =>
+      _beneficiariosController.stream;
+  Stream<List<BeneficiariosGeneral>>
+      get beneficiariosAfiliadosControllerStream =>
+          _beneficiariosAfiliadosController.stream;
 
   dispose() {
-    _beneficiariosController?.close();
-    _beneficiariosAfiliadosController?.close();
+    _beneficiariosController.close();
+    _beneficiariosAfiliadosController.close();
   }
 
   void obtenerBeneficiariosPorIdPersona(String idPersona) async {
@@ -24,14 +29,14 @@ class BeneficiariosBloc {
     _beneficiariosController.sink.add(await benefis(idPersona));
   }
 
-   void obtenerBeneficiariosAfiliadosPorIdPersona(String idPersona) async {
+  void obtenerBeneficiariosAfiliadosPorIdPersona(String idPersona) async {
     _beneficiariosAfiliadosController.sink.add(await benefis(idPersona));
     await beneficiariosApi.obtenerBeneficiarios(idPersona);
     _beneficiariosAfiliadosController.sink.add(await benefis(idPersona));
   }
 
   Future<List<BeneficiariosGeneral>> benefis(String idPersona) async {
-    final List<BeneficiariosGeneral> listBeneficiariosGeneral = []; 
+    final List<BeneficiariosGeneral> listBeneficiariosGeneral = [];
 
     final todosBeneficiarios =
         await beneficiariosDatabase.cargarBeneficiariosPorIdPersona(idPersona);
@@ -108,11 +113,11 @@ class BeneficiariosBloc {
         }
       }
 
-      
       if (listBeneficioFallecido.length > 0) {
         BeneficiariosGeneral beneficiariosGeneral = BeneficiariosGeneral();
         beneficiariosGeneral.tipo = '1';
-        beneficiariosGeneral.texto = 'Beneficiarios con derecho al fondo de cesantía y bolsa de jubulación';
+        beneficiariosGeneral.texto =
+            'Beneficiarios con derecho al fondo de cesantía y bolsa de jubulación';
 
         final List<Beneficiarios> beneAlgo = [];
 
@@ -120,19 +125,23 @@ class BeneficiariosBloc {
           Beneficiarios bene = Beneficiarios();
           bene.nombre = listBeneficioFallecido[i].nombre;
           bene.parentesco = listBeneficioFallecido[i].gradoParentesco;
-          bene.porcentaje = double.parse(listBeneficioFallecido[i].porcentaje).toInt().toString();
+          bene.porcentaje =
+              double.parse(listBeneficioFallecido[i].porcentaje ?? '')
+                  .toInt()
+                  .toString();
           bene.observacion = listBeneficioFallecido[i].observaciones;
 
           beneAlgo.add(bene);
         }
-        beneficiariosGeneral.beneficiarios =beneAlgo;
+        beneficiariosGeneral.beneficiarios = beneAlgo;
         listBeneficiariosGeneral.add(beneficiariosGeneral);
       }
 
       if (listFallecido.length > 0) {
         BeneficiariosGeneral beneficiariosGeneral = BeneficiariosGeneral();
         beneficiariosGeneral.tipo = '0';
-        beneficiariosGeneral.texto = 'Derecho como afiliado al cobro del fondo de funerales';
+        beneficiariosGeneral.texto =
+            'Derecho como afiliado al cobro del fondo de funerales';
 
         final List<Beneficiarios> beneAlgo = [];
 
@@ -140,19 +149,22 @@ class BeneficiariosBloc {
           Beneficiarios bene = Beneficiarios();
           bene.nombre = listFallecido[i].nombre;
           bene.parentesco = listFallecido[i].gradoParentesco;
-          bene.porcentaje = double.parse(listFallecido[i].porcentaje).toInt().toString();
+          bene.porcentaje = double.parse(listFallecido[i].porcentaje ?? '')
+              .toInt()
+              .toString();
           bene.observacion = listFallecido[i].observaciones;
 
           beneAlgo.add(bene);
         }
-        beneficiariosGeneral.beneficiarios =beneAlgo;
+        beneficiariosGeneral.beneficiarios = beneAlgo;
         listBeneficiariosGeneral.add(beneficiariosGeneral);
       }
 
       if (listFuneral.length > 0) {
         BeneficiariosGeneral beneficiariosGeneral = BeneficiariosGeneral();
         beneficiariosGeneral.tipo = '0';
-        beneficiariosGeneral.texto = 'A mi fallecimiento el fondo de funerales debe entregarse a:';
+        beneficiariosGeneral.texto =
+            'A mi fallecimiento el fondo de funerales debe entregarse a:';
 
         final List<Beneficiarios> beneAlgo = [];
 
@@ -160,12 +172,13 @@ class BeneficiariosBloc {
           Beneficiarios bene = Beneficiarios();
           bene.nombre = listFuneral[i].nombre;
           bene.parentesco = listFuneral[i].gradoParentesco;
-          bene.porcentaje = double.parse(listFuneral[i].porcentaje).toInt().toString();
+          bene.porcentaje =
+              double.parse(listFuneral[i].porcentaje ?? '').toInt().toString();
           bene.observacion = listFallecido[i].observaciones;
 
           beneAlgo.add(bene);
         }
-        beneficiariosGeneral.beneficiarios =beneAlgo;
+        beneficiariosGeneral.beneficiarios = beneAlgo;
         listBeneficiariosGeneral.add(beneficiariosGeneral);
       }
 
@@ -182,28 +195,28 @@ class BeneficiariosBloc {
 }
 
 class BeneficiariosList {
-  List<BeneficiariosModel> funeral;
-  List<BeneficiariosModel> fallecido;
-  List<BeneficiariosModel> beneficioFallecido;
-  List<BeneficiariosGeneral> beneficiariosGeneral;
+  List<BeneficiariosModel>? funeral;
+  List<BeneficiariosModel>? fallecido;
+  List<BeneficiariosModel>? beneficioFallecido;
+  List<BeneficiariosGeneral>? beneficiariosGeneral;
 
   BeneficiariosList({this.funeral, this.fallecido, this.beneficioFallecido});
 }
 
 class BeneficiariosGeneral {
-  String tipo;
-  String texto;
-  List<Beneficiarios> beneficiarios;
+  String? tipo;
+  String? texto;
+  List<Beneficiarios>? beneficiarios;
 
   BeneficiariosGeneral({this.tipo, this.beneficiarios, this.texto});
 }
 
 class Beneficiarios {
-  Beneficiarios({this.porcentaje, this.nombre, this.parentesco,this.observacion});
+  Beneficiarios(
+      {this.porcentaje, this.nombre, this.parentesco, this.observacion});
 
-  String porcentaje;
-  String nombre;
-  String parentesco;
-  String observacion;
-  
+  String? porcentaje;
+  String? nombre;
+  String? parentesco;
+  String? observacion;
 }
