@@ -1,24 +1,30 @@
 import 'package:flutter/material.dart';
 
 class SliverCustomHeaderDelegate extends SliverPersistentHeaderDelegate {
-  SliverCustomHeaderDelegate({@required this.minHeight, @required this.maxHeight, @required this.child});
-  final double minHeight;
-  final double maxHeight;
-  final Widget child;
+  SliverCustomHeaderDelegate(
+      {@required this.minHeight,
+      @required this.maxHeight,
+      @required this.child});
+  final double? minHeight;
+  final double? maxHeight;
+  final Widget? child;
 
   @override
-  Widget build(BuildContext context, double shrinkOffset, bool overlapsContent) {
+  Widget build(
+      BuildContext context, double shrinkOffset, bool overlapsContent) {
     return SizedBox.expand(child: child);
   }
 
   @override
-  double get maxExtent => maxHeight;
+  double get maxExtent => maxHeight ?? 0;
 
   @override
-  double get minExtent => minHeight;
+  double get minExtent => minHeight ?? 0;
 
   @override
   bool shouldRebuild(SliverCustomHeaderDelegate oldDelegate) {
-    return maxHeight != oldDelegate.maxHeight || minHeight != oldDelegate.minHeight || child != oldDelegate.child;
+    return maxHeight != oldDelegate.maxHeight ||
+        minHeight != oldDelegate.minHeight ||
+        child != oldDelegate.child;
   }
 }
