@@ -1,36 +1,30 @@
-
-
-
-
-
 import 'dart:convert';
-  
+
 import 'package:http/http.dart' as http;
 import 'package:secedo_unap/src/api/cuotas_prestamo_api.dart';
 import 'package:secedo_unap/src/database/prestamos_database.dart';
 import 'package:secedo_unap/src/model/prestamos_model.dart';
 import 'package:secedo_unap/src/preferencias/preferencias_usuario.dart';
-import 'package:secedo_unap/src/utils/constants.dart'; 
- 
+import 'package:secedo_unap/src/utils/constants.dart';
 
-class PrestamosApi { 
+class PrestamosApi {
   final prestamosDatabase = PrestamosDatabase();
   final cuotasPrestamosApi = CuotasPrestamosApi();
   final preferences = Preferences();
 
   Future<bool> obtenerPrestamos() async {
     try {
-
       //id es el id del usuario
-      final url = Uri.parse('$apiBaseURL/api/Prestamoes/${preferences.idPersona}');
+      final url =
+          Uri.parse('$apiBaseURL/api/Prestamoes/${preferences.idPersona}');
 
-      final resp = await http.get(url);//(url, body: {'id_empresa': id, 'app': 'true', 'tn': prefs.token});
+      final resp = await http.get(
+          url); //(url, body: {'id_empresa': id, 'app': 'true', 'tn': prefs.token});
 
       final decodedData = json.decode(resp.body);
 
       if (decodedData.length > 0) {
         for (int i = 0; i < decodedData.length; i++) {
-
           final date = DateTime.now();
           String dia = date.day.toString();
           String mes = date.month.toString();
@@ -54,8 +48,8 @@ class PrestamosApi {
 
           await prestamosDatabase.insertarPrestamos(prestamosModel);
 
-
-          await cuotasPrestamosApi.obtenerCuotasPrestamos(prestamosModel.idPrestamo);
+          await cuotasPrestamosApi
+              .obtenerCuotasPrestamos(prestamosModel.idPrestamo ?? '');
         }
 
         return true;
