@@ -1,23 +1,24 @@
-
 import 'dart:async';
 
 import 'package:rxdart/rxdart.dart';
 import 'package:secedo_unap/src/api/login_api.dart';
 
-class LoginBloc with Validators {
-
+class LoginBloc extends Validators {
   final loginProviders = LoginApi();
 
-final _emailController = BehaviorSubject<String>();
+  final _emailController = BehaviorSubject<String>();
   final _passwordController = BehaviorSubject<String>();
   final _cargandoLoginController = new BehaviorSubject<bool>();
 
   //Recuperaer los datos del Stream
-  Stream<String> get emailStream =>_emailController.stream.transform(validarname);
-  Stream<String> get passwordStream =>_passwordController.stream.transform(validarPassword);
+  Stream<String> get emailStream =>
+      _emailController.stream.transform(validarname);
+  Stream<String> get passwordStream =>
+      _passwordController.stream.transform(validarPassword);
   Stream<bool> get cargando => _cargandoLoginController.stream;
 
-  Stream<bool> get formValidStream =>Rx.combineLatest2(emailStream, passwordStream, (e, p) => true);
+  Stream<bool> get formValidStream =>
+      Rx.combineLatest2(emailStream, passwordStream, (e, p) => true);
 
   //inserta valores al Stream
   Function(String) get changeEmail => _emailController.sink.add;
@@ -28,15 +29,12 @@ final _emailController = BehaviorSubject<String>();
   String get email => _emailController.value;
   String get password => _passwordController.value;
 
-
-
   dispose() {
-    _emailController?.close();
-    _passwordController?.close();
-    _cargandoLoginController?.close();
+    _emailController.close();
+    _passwordController.close();
+    _cargandoLoginController.close();
   }
 
-  
   Future<bool> login(String user, String pass) async {
     _cargandoLoginController.sink.add(true);
     final resp = await loginProviders.login(email, pass);
@@ -44,7 +42,6 @@ final _emailController = BehaviorSubject<String>();
 
     return resp;
   }
-
 }
 
 class Validators {
@@ -52,7 +49,7 @@ class Validators {
       StreamTransformer<String, String>.fromHandlers(handleData: (email, sink) {
     Pattern pattern =
         r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$';
-    RegExp regExp = new RegExp(pattern);
+    RegExp regExp = new RegExp(pattern.toString());
 
     if (regExp.hasMatch(email)) {
       sink.add(email);
@@ -78,5 +75,4 @@ class Validators {
       sink.addError('Este campo no debe estar vacio');
     }
   });
- 
 }
