@@ -1,51 +1,32 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 import 'package:secedo_unap/src/model/beneficiarios_model.dart';
 import 'database_provider.dart';
 
-class BeneficiariosDatabase{
+class BeneficiariosDatabase {
+  final dbprovider = DatabaseProvider.instance;
 
-  final dbprovider = DatabaseProvider.db;
-
-  insertarBeneficiarios(BeneficiariosModel beneficiarios)async{
-    try{
+  insertarBeneficiarios(BeneficiariosModel beneficiarios) async {
+    try {
       final db = await dbprovider.database;
 
       final res = await db.rawInsert(
           "INSERT OR REPLACE INTO Beneficiarios (idBeneficiario,idPersona,nombre,"
           "gradoParentesco,porcentaje,funeral,beneficioFallecimiento,fallecido,"
           "idGradoParentesco,defuncion,cartaDeclaratoria,observaciones) "
-              "VALUES ('${beneficiarios.idBeneficiario}','${beneficiarios.idPersona}','${beneficiarios.nombre}',"
-              "'${beneficiarios.gradoParentesco}','${beneficiarios.porcentaje}','${beneficiarios.funeral}',"
-              "'${beneficiarios.beneficioFallecimiento}','${beneficiarios.fallecido}','${beneficiarios.idGradoParentesco}',"
-              "'${beneficiarios.defuncion}','${beneficiarios.cartaDeclaratoria}','${beneficiarios.observaciones}')");
+          "VALUES ('${beneficiarios.idBeneficiario}','${beneficiarios.idPersona}','${beneficiarios.nombre}',"
+          "'${beneficiarios.gradoParentesco}','${beneficiarios.porcentaje}','${beneficiarios.funeral}',"
+          "'${beneficiarios.beneficioFallecimiento}','${beneficiarios.fallecido}','${beneficiarios.idGradoParentesco}',"
+          "'${beneficiarios.defuncion}','${beneficiarios.cartaDeclaratoria}','${beneficiarios.observaciones}')");
       return res;
-
-    }catch(exception){
+    } catch (exception) {
       print(exception);
     }
   }
 
-  Future<List<BeneficiariosModel>> cargarBeneficiariosPorIdPersona(String idPersona) async {
-    final db = await dbprovider.database; 
-    final res =
-    await db.rawQuery("SELECT * FROM Beneficiarios where idPersona='$idPersona'");
+  Future<List<BeneficiariosModel>> cargarBeneficiariosPorIdPersona(
+      String idPersona) async {
+    final db = await dbprovider.database;
+    final res = await db
+        .rawQuery("SELECT * FROM Beneficiarios where idPersona='$idPersona'");
 
     List<BeneficiariosModel> list = res.isNotEmpty
         ? res.map((c) => BeneficiariosModel.fromJson(c)).toList()
@@ -53,9 +34,6 @@ class BeneficiariosDatabase{
 
     return list;
   }
- 
-
-
 
   deleteBeneficiarios() async {
     final db = await dbprovider.database;
@@ -64,6 +42,4 @@ class BeneficiariosDatabase{
 
     return res;
   }
-
- 
 }
