@@ -1,54 +1,35 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 import 'package:secedo_unap/src/model/descuento_planilla_model.dart';
 
 import 'database_provider.dart';
 
-class DescuentoPlanillaDatabase{
+class DescuentoPlanillaDatabase {
+  final dbprovider = DatabaseProvider.instance;
 
-  final dbprovider = DatabaseProvider.db;
-
-  insertarDescuentosPlanillas(DescuentoPlanillaModel descuentoPlanillaModel)async{
-    try{
+  insertarDescuentosPlanillas(
+      DescuentoPlanillaModel descuentoPlanillaModel) async {
+    try {
       final db = await dbprovider.database;
 
       final res = await db.rawInsert(
           "INSERT OR REPLACE INTO DescuentoPlanilla (idDescuentoPlanilla,idAfiliacion,codigo,"
           "nombre,fecha,cesantia,funeral,jubilacion,otros,multa,apr,garantizado,"
           "descuento,ntotal,aplicado,diferencia,idPersona,posicion) "
-              "VALUES ('${descuentoPlanillaModel.idDescuentoPlanilla}','${descuentoPlanillaModel.idAfiliacion}',"
-              "'${descuentoPlanillaModel.codigo}','${descuentoPlanillaModel.nombre}','${descuentoPlanillaModel.fecha}',"
-              "'${descuentoPlanillaModel.cesantia}','${descuentoPlanillaModel.funeral}','${descuentoPlanillaModel.jubilacion}',"
-              "'${descuentoPlanillaModel.otros}','${descuentoPlanillaModel.multa}','${descuentoPlanillaModel.apr}',"
-              "'${descuentoPlanillaModel.garantizado}','${descuentoPlanillaModel.descuento}','${descuentoPlanillaModel.ntotal}',"
-              "'${descuentoPlanillaModel.aplicado}','${descuentoPlanillaModel.diferencia}','${descuentoPlanillaModel.idPersona}','${descuentoPlanillaModel.posicion}')");
+          "VALUES ('${descuentoPlanillaModel.idDescuentoPlanilla}','${descuentoPlanillaModel.idAfiliacion}',"
+          "'${descuentoPlanillaModel.codigo}','${descuentoPlanillaModel.nombre}','${descuentoPlanillaModel.fecha}',"
+          "'${descuentoPlanillaModel.cesantia}','${descuentoPlanillaModel.funeral}','${descuentoPlanillaModel.jubilacion}',"
+          "'${descuentoPlanillaModel.otros}','${descuentoPlanillaModel.multa}','${descuentoPlanillaModel.apr}',"
+          "'${descuentoPlanillaModel.garantizado}','${descuentoPlanillaModel.descuento}','${descuentoPlanillaModel.ntotal}',"
+          "'${descuentoPlanillaModel.aplicado}','${descuentoPlanillaModel.diferencia}','${descuentoPlanillaModel.idPersona}','${descuentoPlanillaModel.posicion}')");
       return res;
-
-    }catch(exception){
+    } catch (exception) {
       print(exception);
     }
   }
 
   Future<List<DescuentoPlanillaModel>> cargarDescuentosPlanillas() async {
-    final db = await dbprovider.database; 
-    final res =
-    await db.rawQuery("SELECT * FROM DescuentoPlanilla order by idDescuentoPlanilla desc");
+    final db = await dbprovider.database;
+    final res = await db.rawQuery(
+        "SELECT * FROM DescuentoPlanilla order by idDescuentoPlanilla desc");
 
     List<DescuentoPlanillaModel> list = res.isNotEmpty
         ? res.map((c) => DescuentoPlanillaModel.fromJson(c)).toList()
@@ -56,7 +37,6 @@ class DescuentoPlanillaDatabase{
 
     return list;
   }
- 
 
   deleteDescuentoPlanilla() async {
     final db = await dbprovider.database;
@@ -65,5 +45,4 @@ class DescuentoPlanillaDatabase{
 
     return res;
   }
- 
 }
