@@ -1,69 +1,37 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-import 'package:secedo_unap/src/model/deudas_model.dart'; 
+import 'package:secedo_unap/src/model/deudas_model.dart';
 
 import 'database_provider.dart';
 
-class DeudasDatabase{
+class DeudasDatabase {
+  final dbprovider = DatabaseProvider.instance;
 
-  final dbprovider = DatabaseProvider.db;
-
-  insertarDeudas(DeudasModel deudas)async{
-    try{
+  insertarDeudas(DeudasModel deudas) async {
+    try {
       final db = await dbprovider.database;
 
       final res = await db.rawInsert(
           "INSERT OR REPLACE INTO Deudas (idPersona,idAfiliacion,codigo,"
           "nombre,cesantia,funeral,jubilacion,multa,apr,capital,interes,garantizado,fechaAtual,descuento) "
-              "VALUES ('${deudas.idPersona}','${deudas.idAfiliacion}',"
-              "'${deudas.codigo}','${deudas.nombre}','${deudas.cesantia}',"
-              "'${deudas.funeral}','${deudas.jubilacion}','${deudas.multa}',"
-              "'${deudas.apr}','${deudas.capital}','${deudas.interes}',"
-              "'${deudas.garantizado}','${deudas.fechaAtual}','${deudas.descuento}')");
+          "VALUES ('${deudas.idPersona}','${deudas.idAfiliacion}',"
+          "'${deudas.codigo}','${deudas.nombre}','${deudas.cesantia}',"
+          "'${deudas.funeral}','${deudas.jubilacion}','${deudas.multa}',"
+          "'${deudas.apr}','${deudas.capital}','${deudas.interes}',"
+          "'${deudas.garantizado}','${deudas.fechaAtual}','${deudas.descuento}')");
       return res;
-
-    }catch(exception){
+    } catch (exception) {
       print(exception);
     }
   }
 
   Future<List<DeudasModel>> cargarDeudas() async {
-    final db = await dbprovider.database; 
-    final res =
-    await db.rawQuery("SELECT * FROM Deudas ");
+    final db = await dbprovider.database;
+    final res = await db.rawQuery("SELECT * FROM Deudas ");
 
-    List<DeudasModel> list = res.isNotEmpty
-        ? res.map((c) => DeudasModel.fromJson(c)).toList()
-        : [];
+    List<DeudasModel> list =
+        res.isNotEmpty ? res.map((c) => DeudasModel.fromJson(c)).toList() : [];
 
     return list;
   }
- 
 
   deleteDeudas() async {
     final db = await dbprovider.database;
@@ -72,5 +40,4 @@ class DeudasDatabase{
 
     return res;
   }
- 
 }
