@@ -1,55 +1,31 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 import 'package:secedo_unap/src/model/planilla_enviada_model.dart';
 
 import 'database_provider.dart';
 
-class PlanillaEnviadaDatabase{
+class PlanillaEnviadaDatabase {
+  final dbprovider = DatabaseProvider.instance;
 
-  final dbprovider = DatabaseProvider.db;
-
-  insertarPlanillaEnviada(PlanillaEnviadaModel planillaEnviadaModel)async{
-    try{
+  insertarPlanillaEnviada(PlanillaEnviadaModel planillaEnviadaModel) async {
+    try {
       final db = await dbprovider.database;
 
       final res = await db.rawInsert(
           "INSERT OR REPLACE INTO PlanillaEnviada (idPlanillaEnviada,idAfiliacion,tipoPlanilla,"
           "fecha,cesantia,funeral,jubilacion,multa,apr,garantizado,descuento,total) "
-              "VALUES ('${planillaEnviadaModel.idPlanillaEnviada}','${planillaEnviadaModel.idAfiliacion}',"
-              "'${planillaEnviadaModel.tipoPlanilla}','${planillaEnviadaModel.fecha}','${planillaEnviadaModel.cesantia}',"
-              "'${planillaEnviadaModel.funeral}','${planillaEnviadaModel.jubilacion}','${planillaEnviadaModel.multa}',"
-              "'${planillaEnviadaModel.apr}','${planillaEnviadaModel.garantizado}','${planillaEnviadaModel.descuento}',"
-              "'${planillaEnviadaModel.total}')");
+          "VALUES ('${planillaEnviadaModel.idPlanillaEnviada}','${planillaEnviadaModel.idAfiliacion}',"
+          "'${planillaEnviadaModel.tipoPlanilla}','${planillaEnviadaModel.fecha}','${planillaEnviadaModel.cesantia}',"
+          "'${planillaEnviadaModel.funeral}','${planillaEnviadaModel.jubilacion}','${planillaEnviadaModel.multa}',"
+          "'${planillaEnviadaModel.apr}','${planillaEnviadaModel.garantizado}','${planillaEnviadaModel.descuento}',"
+          "'${planillaEnviadaModel.total}')");
       return res;
-
-    }catch(exception){
+    } catch (exception) {
       print(exception);
     }
   }
 
   Future<List<PlanillaEnviadaModel>> cargarPlanillaEnviada() async {
-    final db = await dbprovider.database; 
-    final res =
-    await db.rawQuery("SELECT * FROM PlanillaEnviada ");
+    final db = await dbprovider.database;
+    final res = await db.rawQuery("SELECT * FROM PlanillaEnviada ");
 
     List<PlanillaEnviadaModel> list = res.isNotEmpty
         ? res.map((c) => PlanillaEnviadaModel.fromJson(c)).toList()
@@ -57,7 +33,6 @@ class PlanillaEnviadaDatabase{
 
     return list;
   }
- 
 
   deletePlanillaEnviada() async {
     final db = await dbprovider.database;
@@ -66,5 +41,4 @@ class PlanillaEnviadaDatabase{
 
     return res;
   }
- 
 }
