@@ -1,47 +1,30 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 import 'package:secedo_unap/src/model/prestamos_model.dart';
 
 import 'database_provider.dart';
 
-class PrestamosDatabase{
+class PrestamosDatabase {
+  final dbprovider = DatabaseProvider.instance;
 
-  final dbprovider = DatabaseProvider.db;
-
-  insertarPrestamos(PrestamosModel prestamos)async{
-    try{
+  insertarPrestamos(PrestamosModel prestamos) async {
+    try {
       final db = await dbprovider.database;
 
       final res = await db.rawInsert(
           "INSERT OR REPLACE INTO Prestamos (idPrestamo,cheque,fSolicitud,"
           "solicitado,fAprobado,aprobado,girar,tInteres,nCuotas,tipo,prioridad,garante,fechaActualizado,idPersona) "
-              "VALUES ('${prestamos.idPrestamo}','${prestamos.cheque}','${prestamos.fSolicitud}',"
-              "'${prestamos.solicitado}','${prestamos.fAprobado}','${prestamos.aprobado}','${prestamos.girar}',"
-              "'${prestamos.tInteres}','${prestamos.nCuotas}','${prestamos.tipo}','${prestamos.prioridad}',"
-              "'${prestamos.garante}','${prestamos.fechaActualizado}','${prestamos.idPersona}')");
+          "VALUES ('${prestamos.idPrestamo}','${prestamos.cheque}','${prestamos.fSolicitud}',"
+          "'${prestamos.solicitado}','${prestamos.fAprobado}','${prestamos.aprobado}','${prestamos.girar}',"
+          "'${prestamos.tInteres}','${prestamos.nCuotas}','${prestamos.tipo}','${prestamos.prioridad}',"
+          "'${prestamos.garante}','${prestamos.fechaActualizado}','${prestamos.idPersona}')");
       return res;
-
-    }catch(exception){
+    } catch (exception) {
       print(exception);
     }
   }
 
   Future<List<PrestamosModel>> cargarPrestamos() async {
-    final db = await dbprovider.database; 
-    final res =
-    await db.rawQuery("SELECT * FROM Prestamos");
+    final db = await dbprovider.database;
+    final res = await db.rawQuery("SELECT * FROM Prestamos");
 
     List<PrestamosModel> list = res.isNotEmpty
         ? res.map((c) => PrestamosModel.fromJson(c)).toList()
@@ -49,8 +32,6 @@ class PrestamosDatabase{
 
     return list;
   }
- 
-
 
   deletePrestamos() async {
     final db = await dbprovider.database;
@@ -59,6 +40,4 @@ class PrestamosDatabase{
 
     return res;
   }
-
- 
 }
